Deprecate cape-ispell in favor of cape-dict
cape-ispell just used grep to scan /usr/share/dict/words. We can replicate
`ispell-lookup-words' with very little code thanks to the function
`process-lines-ignore-status'. This way we avoid loading ispell.el. Also the
name `cape-ispell' was rather confusing since it suggested that some kind of
spell-checking is going on, which was never the case.