char: Define `*--ensure-*` functions as macros
This fixes the void-function errors. Efficiency wise IMO it's pretty
much on par with the function implementation. But when the current
translation function having `eval-when-compile` around it, this
suppresses the cape-char--translation-hash symbol definition is void
errors.

There are possibly more targeted fixes to this problem, and we'll
probably be changing the `eval-when-compile` in the future anyway, but I
think this change isn't really a "hack" to fix it -- just an alternate
solution for `*--ensure-*` functions (macros).