cape-wrap-inside-faces: use (1- (point)) instead of (point)
This addresses #109 where we weren't picking up
`tree-sitter-hl-face:comment`. The problem is that when you
use (text-properties-at (point)), it returns the properties of the
character after the point, not the character before the point. To
illustrate, running `(describe-char (point-max))` errors out with "No
character follows specified position". When we are completing based on
face, we always mean the face of the character immediately before the
point.